Move slothandler and json to liboxide
